refactor some code

+++ alicloud/service_alicloud_flink_job.go
@@ -19,18 +19,31 @@ func parseJobId(id string) (string, string, string, error) {
 
 // Job methods
 func (s *FlinkService) DescribeFlinkJob(id string) (*aliyunFlinkAPI.Job, error) {
+	// Validate input parameter
+	if id == "" {
+		return nil, fmt.Errorf("job ID cannot be empty")
+	}
+
 	// Parse job ID to extract workspace ID, namespace and job ID
 	// Format: workspaceId:namespace:jobId
 	workspaceId, namespaceName, jobId, err := parseJobId(id)
 	if err != nil {
-		return nil, err
+		return nil, WrapErrorf(err, DefaultErrorMsg, id, "parseJobId", AlibabaCloudSdkGoERROR)
 	}
 
-	return s.flinkAPI.GetJob(workspaceId, namespaceName, jobId)
+	job, err := s.flinkAPI.GetJob(workspaceId, namespaceName, jobId)
+	if err != nil {
+		return nil, WrapErrorf(err, DefaultErrorMsg, id, "GetJob", AlibabaCloudSdkGoERROR)
+	}
+
+	return job, nil
 }
 
 func (s *FlinkService) StartJob(params *aliyunFlinkAPI.JobStartParameters) (*aliyunFlinkAPI.Job, error) {
 	// Validate required parameters
+	if params == nil {
+		return nil, fmt.Errorf("JobStartParameters cannot be nil")
+	}
 	if params.WorkspaceId == "" {
 		return nil, fmt.Errorf("WorkspaceId is required in JobStartParameters")
 	}
@@ -38,23 +51,46 @@ func (s *FlinkService) StartJob(params *aliyunFlinkAPI.JobStartParameters) (*ali
 		return nil, fmt.Errorf("Namespace is required in JobStartParameters")
 	}
 
-	return s.flinkAPI.StartJob(params)
+	job, err := s.flinkAPI.StartJob(params)
+	if err != nil {
+		return nil, WrapErrorf(err, DefaultErrorMsg, fmt.Sprintf("%s:%s", params.WorkspaceId, params.Namespace), "StartJob", AlibabaCloudSdkGoERROR)
+	}
+
+	return job, nil
 }
 
 func (s *FlinkService) UpdateJob(stateId string, updateParams *aliyunFlinkAPI.HotUpdateJobParams) (*aliyunFlinkAPI.HotUpdateJobResult, error) {
+	// Validate input parameters
+	if stateId == "" {
+		return nil, fmt.Errorf("job state ID cannot be empty")
+	}
+	if updateParams == nil {
+		return nil, fmt.Errorf("update parameters cannot be nil")
+	}
+
 	// Parse job ID to extract workspace ID, namespace and job ID
 	workspaceId, namespaceName, jobId, err := parseJobId(stateId)
 	if err != nil {
-		return nil, err
+		return nil, WrapErrorf(err, DefaultErrorMsg, stateId, "parseJobId", AlibabaCloudSdkGoERROR)
 	}
 
-	return s.flinkAPI.UpdateJob(workspaceId, namespaceName, jobId, updateParams)
+	result, err := s.flinkAPI.UpdateJob(workspaceId, namespaceName, jobId, updateParams)
+	if err != nil {
+		return nil, WrapErrorf(err, DefaultErrorMsg, stateId, "UpdateJob", AlibabaCloudSdkGoERROR)
+	}
+
+	return result, nil
 }
 
 func (s *FlinkService) StopJob(stateId string, withSavepoint bool) error {
+	// Validate input parameter
+	if stateId == "" {
+		return fmt.Errorf("job state ID cannot be empty")
+	}
+
 	workspaceId, namespaceName, jobId, err := parseJobId(stateId)
 	if err != nil {
-		return err
+		return WrapErrorf(err, DefaultErrorMsg, stateId, "parseJobId", AlibabaCloudSdkGoERROR)
 	}
 
 	// First get the current job status
@@ -64,13 +100,18 @@ func (s *FlinkService) StopJob(stateId string, withSavepoint bool) error {
 			// Job doesn't exist, no need to stop
 			return nil
 		}
-		return err
+		return WrapErrorf(err, DefaultErrorMsg, stateId, "GetJob", AlibabaCloudSdkGoERROR)
 	}
 
 	// Check if job is running, only stop if it's running
-	jobStatus := job.GetStatus()
-	if jobStatus == aliyunFlinkAPI.FlinkJobStatusRunning.String() {
-		return s.flinkAPI.StopJob(workspaceId, namespaceName, jobId, withSavepoint)
+	if job != nil {
+		jobStatus := job.GetStatus()
+		if jobStatus == aliyunFlinkAPI.FlinkJobStatusRunning.String() {
+			err = s.flinkAPI.StopJob(workspaceId, namespaceName, jobId, withSavepoint)
+			if err != nil {
+				return WrapErrorf(err, DefaultErrorMsg, stateId, "StopJob", AlibabaCloudSdkGoERROR)
+			}
+		}
 	}
 
 	// Job is not running, no need to stop
@@ -78,15 +119,39 @@ func (s *FlinkService) StopJob(stateId string, withSavepoint bool) error {
 }
 
 func (s *FlinkService) DeleteJob(stateId string) error {
+	// Validate input parameter
+	if stateId == "" {
+		return fmt.Errorf("job state ID cannot be empty")
+	}
+
 	workspaceId, namespaceName, jobId, err := parseJobId(stateId)
 	if err != nil {
-		return err
+		return WrapErrorf(err, DefaultErrorMsg, stateId, "parseJobId", AlibabaCloudSdkGoERROR)
 	}
-	return s.flinkAPI.DeleteJob(workspaceId, namespaceName, jobId)
+
+	err = s.flinkAPI.DeleteJob(workspaceId, namespaceName, jobId)
+	if err != nil {
+		return WrapErrorf(err, DefaultErrorMsg, stateId, "DeleteJob", AlibabaCloudSdkGoERROR)
+	}
+
+	return nil
 }
 
 func (s *FlinkService) ListJobs(workspaceId, namespaceName, deploymentId string) ([]aliyunFlinkAPI.Job, error) {
-	return s.flinkAPI.ListJobs(workspaceId, namespaceName, deploymentId)
+	// Validate required parameters
+	if workspaceId == "" {
+		return nil, fmt.Errorf("workspaceId cannot be empty")
+	}
+	if namespaceName == "" {
+		return nil, fmt.Errorf("namespaceName cannot be empty")
+	}
+
+	jobs, err := s.flinkAPI.ListJobs(workspaceId, namespaceName, deploymentId)
+	if err != nil {
+		return nil, WrapErrorf(err, DefaultErrorMsg, fmt.Sprintf("%s:%s", workspaceId, namespaceName), "ListJobs", AlibabaCloudSdkGoERROR)
+	}
+
+	return jobs, nil
 }
 
 func (s *FlinkService) FlinkJobStateRefreshFunc(id string, failStates []string) resource.StateRefreshFunc {
